fix: move all tests to src/__tests__ directory
