+++ src/__tests__/pages/workflow/WorkflowIndex/WorkflowIndexLoading.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { WorkflowIndexLoading } from './WorkflowIndexLoading';
+import { WorkflowIndexLoading } from '../../../../pages/workflow/WorkflowIndex/WorkflowIndexLoading';
 
 /**
  * Mock: display hook
